remove extra line breaks in text field
