+++ app/javascript/stylesheets/application.scss
@@ -4,3 +4,4 @@ $link-decoration: none;
 @import 'bootstrap/scss/bootstrap';
 @import 'select2';
 @import 'select2-bootstrap-5-theme';
+@import 'bootstrap-icons/font/bootstrap-icons';
